+++ src/theme/components/link.ts
@@ -13,9 +13,6 @@ export const baseLinkStyles = {
   _hover: {
     color: "ui.link.secondary",
   },
-  _visited: {
-    color: "ui.link.tertiary",
-  },
 };
 
 const baseButtonLinkStyles = {
@@ -42,11 +39,6 @@ const variants = {
       color: "ui.link.secondary",
       textDecoration: "underline",
     },
-    _visited: {
-      svg: {
-        color: "ui.link.tertiary",
-      },
-    },
   },
   // The "button" variant is deprecated.
   button: {
@@ -66,9 +58,6 @@ const variants = {
       color: "ui.white",
       bg: "ui.link.secondary",
       textDecoration: "none",
-    },
-    _visited: {
-      color: "ui.white",
     },
   },
   buttonPrimary: {